requirements.txt
Django==1.8.4
Logbook==0.9.1
MySQL-python==1.2.5
Pillow==2.5.1
PyJWT==1.4.0
argparse==1.3.0
beautifulsoup4==4.3.2
cffi==0.9.2
cryptography==0.8.2
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-classy-tags==0.6.2
django-crispy-forms==1.4.0
django-easy-maps==0.9.2
django-geojson==2.8.1
django-geoposition==0.2.2
django-leaflet==0.16.0
django-realtime==1.1
django-registration-redux==1.2
django-tinymce==2.0.2
django-toolbelt==0.0.1
dnet==1.12
enum34==1.0.4
foreman==0.9.7
futures==3.0.3
geopy==1.11.0
gunicorn==19.3.0
jsonfield==1.0.3
lxml==3.4.4
ndg-httpsclient==0.3.3
numpy==1.9.0
oauthlib==0.7.2
pbr==1.6.0
psycopg2==2.6.1
py==1.4.30
pyOpenSSL==0.15.1
pyasn1==0.1.7
pycparser==2.12
pyshark==0.3.5
python-openid==2.2.5
python-social-auth==0.2.12
pytz==2015.4
requests==2.6.2
requests-futures==0.9.5
requests-oauthlib==0.4.2
selenium==2.45.0
simplejson==3.8.0
six==1.9.0
static3==0.6.1
stevedore==1.7.0
trollius==1.0.4
tweepy==3.3.0
vboxapi==1.0
virtualenv==13.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.0
wsgiref==0.1.2
